Reactoring and tested in arm (rasperry pi)

+++ src/effects/BlinnPhongEffect.hpp
@@ -104,7 +104,7 @@ public:
         
             // Specular component: spec = (N · H)^shininess
             float specAngle = std::max(0.0f, smath::dot(N,scene.halfwayVector)); // viewer
-            float spec = std::powf(specAngle, tri.material.Ns); // Blinn Phong shininess needs *4 to be like Phong
+            float spec = std::pow(specAngle, tri.material.Ns); // Blinn Phong shininess needs *4 to be like Phong
         
             slib::vec3 color = Ka + Kd * diff + Ks * spec;
             return Color(color).toBgra();
