+++ src/effects/PhongEffect.hpp
@@ -99,7 +99,7 @@ public:
         
             slib::vec3 R = smath::normalize(normal * 2.0f * smath::dot(normal,scene.lux) - scene.lux);
             float specAngle = std::max(0.0f, smath::dot(R,scene.eye)); // viewer
-            float spec = std::powf(specAngle, tri.material.Ns);
+            float spec = std::pow(specAngle, tri.material.Ns);
         
             slib::vec3 color = Ka + Kd * diff + Ks * spec;
             return Color(color).toBgra(); // assumes vec3 uses .r/g/b or [0]/[1]/[2]
