+++ src/effects/TexturedBlinnPhongEffect.hpp
@@ -111,7 +111,7 @@ public:
         
             // Specular component: spec = (N · H)^shininess
             float specAngle = std::max(0.0f, smath::dot(N,scene.halfwayVector)); // viewer
-            float spec = std::powf(specAngle, tri.material.Ns); // Blinn Phong shininess needs *4 to be like Phong
+            float spec = std::pow(specAngle, tri.material.Ns); // Blinn Phong shininess needs *4 to be like Phong
         
             float w = 1 / vRaster.tex.w;
             if (tri.material.map_Kd.textureFilter == slib::TextureFilter::BILINEAR) {
